Move healing code to state
