+++ src/Nethermind/Nethermind.State/IWorldStateManager.cs
@@ -2,7 +2,11 @@
 // SPDX-License-Identifier: LGPL-3.0-only
 
 using System;
+using System.Collections.Generic;
 using Nethermind.Core;
+using Nethermind.Core.Crypto;
+using Nethermind.State.Healing;
+using Nethermind.State.Snap;
 using Nethermind.State.SnapServer;
 using Nethermind.Trie.Pruning;
 
@@ -26,8 +30,10 @@ public interface IWorldStateManager
 
     // TODO: These two method can be combined
     IOverridableWorldScope CreateOverridableWorldScope();
-
     IWorldState CreateOverlayWorldState(IKeyValueStoreWithBatching overlayState, IKeyValueStore overlayCode);
+
+
+    void InitializeNetwork(ITrieNodeRecovery<IReadOnlyList<Hash256>> hashRecovery, ITrieNodeRecovery<GetTrieNodesRequest> nodeRecovery);
 }
 
 public interface IOverridableWorldScope
