+++ src/Nethermind/Nethermind.Synchronization.Test/Trie/HealingTreeTests.cs
@@ -10,6 +10,7 @@ using Nethermind.Core.Extensions;
 using Nethermind.Core.Test;
 using Nethermind.Core.Test.Builders;
 using Nethermind.Logging;
+using Nethermind.State.Healing;
 using Nethermind.State.Snap;
 using Nethermind.Synchronization.Trie;
 using Nethermind.Trie;
