+++ src/Nethermind/Nethermind.Synchronization.Test/Trie/RecoveryTests.cs
@@ -12,6 +12,7 @@ using Nethermind.Core.Crypto;
 using Nethermind.Core.Test.Builders;
 using Nethermind.Logging;
 using Nethermind.Network.Contract.P2P;
+using Nethermind.State.Healing;
 using Nethermind.State.Snap;
 using Nethermind.Synchronization.Peers;
 using Nethermind.Synchronization.Trie;
